Refactoring tests for testing internal components without exporting them.
