+++ docs/implementors/core/convert/trait.From.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {};
-implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/convert/trait.From.html' title='core::convert::From'>From</a>&lt;<a class='struct' href='rusty_secrets/custom_error/struct.Error.html' title='rusty_secrets::custom_error::Error'>Error</a>&gt; for <a class='struct' href='https://doc.rust-lang.org/nightly/std/io/error/struct.Error.html' title='std::io::error::Error'>Error</a>",];
+implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/convert/trait.From.html' title='core::convert::From'>From</a>&lt;Error&gt; for <a class='struct' href='https://doc.rust-lang.org/nightly/std/io/error/struct.Error.html' title='std::io::error::Error'>Error</a>",];
 
             if (window.register_implementors) {
                 window.register_implementors(implementors);
