+++ docs/implementors/core/fmt/trait.Debug.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {};
-implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html' title='core::fmt::Debug'>Debug</a> for <a class='struct' href='rusty_secrets/custom_error/struct.Error.html' title='rusty_secrets::custom_error::Error'>Error</a>","impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html' title='core::fmt::Debug'>Debug</a> for <a class='struct' href='rusty_secrets/share_data/struct.ShareData.html' title='rusty_secrets::share_data::ShareData'>ShareData</a>",];
+implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/fmt/trait.Debug.html' title='core::fmt::Debug'>Debug</a> for <a class='struct' href='rusty_secrets/share_data/struct.ShareData.html' title='rusty_secrets::share_data::ShareData'>ShareData</a>",];
 
             if (window.register_implementors) {
                 window.register_implementors(implementors);
