--- docs/implementors/core/fmt/trait.Display.js
@@ -1,10 +1,0 @@
-(function() {var implementors = {};
-implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/core/fmt/trait.Display.html' title='core::fmt::Display'>Display</a> for <a class='struct' href='rusty_secrets/custom_error/struct.Error.html' title='rusty_secrets::custom_error::Error'>Error</a>",];
-
-            if (window.register_implementors) {
-                window.register_implementors(implementors);
-            } else {
-                window.pending_implementors = implementors;
-            }
-        
-})()
