--- docs/implementors/std/error/trait.Error.js
@@ -1,10 +1,0 @@
-(function() {var implementors = {};
-implementors["rusty_secrets"] = ["impl <a class='trait' href='https://doc.rust-lang.org/nightly/std/error/trait.Error.html' title='std::error::Error'>Error</a> for <a class='struct' href='rusty_secrets/custom_error/struct.Error.html' title='rusty_secrets::custom_error::Error'>Error</a>",];
-
-            if (window.register_implementors) {
-                window.register_implementors(implementors);
-            } else {
-                window.pending_implementors = implementors;
-            }
-        
-})()
