--- docs/rusty_secrets/custom_error/Error.t.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta http-equiv="refresh" content="0;URL=struct.Error.html">
-</head>
-<body>
-    <p>Redirecting to <a href="struct.Error.html">struct.Error.html</a>...</p>
-    <script>location.replace("struct.Error.html" + location.search + location.hash);</script>
-</body>
-</html>
