--- docs/rusty_secrets/custom_error/other_io_err.v.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta http-equiv="refresh" content="0;URL=fn.other_io_err.html">
-</head>
-<body>
-    <p>Redirecting to <a href="fn.other_io_err.html">fn.other_io_err.html</a>...</p>
-    <script>location.replace("fn.other_io_err.html" + location.search + location.hash);</script>
-</body>
-</html>
