--- docs/rusty_secrets/custom_error/pie2io.v.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta http-equiv="refresh" content="0;URL=fn.pie2io.html">
-</head>
-<body>
-    <p>Redirecting to <a href="fn.pie2io.html">fn.pie2io.html</a>...</p>
-    <script>location.replace("fn.pie2io.html" + location.search + location.hash);</script>
-</body>
-</html>
