--- docs/rusty_secrets/custom_error/sidebar-items.js
@@ -1,1 +1,0 @@
-initSidebarItems({"fn":[["other_io_err","Returns an `io::Error` from description string and optional detail string. Particularly useful in `Result` expressions."],["pie2io","maps a `ParseIntError` to an `io::Error`"]],"struct":[["Error","Error struct used for generating an `io::Error` from a generic description."]]});
