+++ docs/rusty_secrets/fn.generate_shares.html
@@ -47,7 +47,7 @@
                    <a id="toggle-all-docs" href="javascript:void(0)" title="collapse all docs">
                        [<span class='inner'>&#x2212;</span>]
                    </a>
-               </span><a id='src-216' class='srclink' href='../src/rusty_secrets/src/sss.rs.html#32-67' title='goto source code'>[src]</a></span></h1>
+               </span><a id='src-228' class='srclink' href='../src/rusty_secrets/src/sss.rs.html#32-67' title='goto source code'>[src]</a></span></h1>
 <pre class='rust fn'>pub fn generate_shares(k: <a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.u8.html'>u8</a>, n: <a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.u8.html'>u8</a>, secret: <a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.slice.html'>&amp;[</a><a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.u8.html'>u8</a><a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.slice.html'>]</a>, sign_shares: <a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.bool.html'>bool</a>) -&gt; <a class='type' href='https://doc.rust-lang.org/nightly/std/io/error/type.Result.html' title='std::io::error::Result'>Result</a>&lt;<a class='struct' href='https://doc.rust-lang.org/nightly/collections/vec/struct.Vec.html' title='collections::vec::Vec'>Vec</a>&lt;<a class='struct' href='https://doc.rust-lang.org/nightly/collections/string/struct.String.html' title='collections::string::String'>String</a>&gt;&gt;</pre><div class='docblock'><p>Performs threshold k-out-of-n Shamir secret sharing.</p>
 
 <h1 id='examples' class='section-header'><a href='#examples'>Examples</a></h1>
