+++ docs/rusty_secrets/fn.recover_secret.html
@@ -47,7 +47,7 @@
                    <a id="toggle-all-docs" href="javascript:void(0)" title="collapse all docs">
                        [<span class='inner'>&#x2212;</span>]
                    </a>
-               </span><a id='src-252' class='srclink' href='../src/rusty_secrets/src/sss.rs.html#111-126' title='goto source code'>[src]</a></span></h1>
+               </span><a id='src-264' class='srclink' href='../src/rusty_secrets/src/sss.rs.html#111-126' title='goto source code'>[src]</a></span></h1>
 <pre class='rust fn'>pub fn recover_secret(shares: <a class='struct' href='https://doc.rust-lang.org/nightly/collections/vec/struct.Vec.html' title='collections::vec::Vec'>Vec</a>&lt;<a class='struct' href='https://doc.rust-lang.org/nightly/collections/string/struct.String.html' title='collections::string::String'>String</a>&gt;, verify_signatures: <a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.bool.html'>bool</a>) -&gt; <a class='type' href='https://doc.rust-lang.org/nightly/std/io/error/type.Result.html' title='std::io::error::Result'>Result</a>&lt;<a class='struct' href='https://doc.rust-lang.org/nightly/collections/vec/struct.Vec.html' title='collections::vec::Vec'>Vec</a>&lt;<a class='primitive' href='https://doc.rust-lang.org/nightly/std/primitive.u8.html'>u8</a>&gt;&gt;</pre><div class='docblock'><p>Recovers the secret from a k-out-of-n Shamir secret sharing.</p>
 
 <p>At least <code>k</code> distinct shares need to be provided to recover the share.</p>
