+++ docs/rusty_secrets/index.html
@@ -47,16 +47,10 @@
                    <a id="toggle-all-docs" href="javascript:void(0)" title="collapse all docs">
                        [<span class='inner'>&#x2212;</span>]
                    </a>
-               </span><a id='src-0' class='srclink' href='../src/rusty_secrets/src/lib.rs.html#1-28' title='goto source code'>[src]</a></span></h1>
-<h2 id='modules' class='section-header'><a href="#modules">Modules</a></h2>
+               </span><a id='src-0' class='srclink' href='../src/rusty_secrets/src/lib.rs.html#1-35' title='goto source code'>[src]</a></span></h1>
+<div class='docblock'><p><code>RustySecrets</code> implements Shamir Secret Sharing in Rust. It provides the possibility to sign shares.</p>
+</div><h2 id='modules' class='section-header'><a href="#modules">Modules</a></h2>
 <table>
-                       <tr class=' module-item'>
-                           <td><a class='mod' href='custom_error/index.html'
-                                  title='rusty_secrets::custom_error'>custom_error</a></td>
-                           <td class='docblock-short'>
-                                
-                           </td>
-                       </tr>
                        <tr class=' module-item'>
                            <td><a class='mod' href='share_data/index.html'
                                   title='rusty_secrets::share_data'>share_data</a></td>
