+++ docs/rusty_secrets/share_data/fn.file_descriptor_proto.html
@@ -47,7 +47,7 @@
                    <a id="toggle-all-docs" href="javascript:void(0)" title="collapse all docs">
                        [<span class='inner'>&#x2212;</span>]
                    </a>
-               </span><a id='src-450' class='srclink' href='../../src/rusty_secrets/src/share_data.rs.html#339-345' title='goto source code'>[src]</a></span></h1>
+               </span><a id='src-452' class='srclink' href='../../src/rusty_secrets/src/share_data.rs.html#339-345' title='goto source code'>[src]</a></span></h1>
 <pre class='rust fn'>pub fn file_descriptor_proto() -&gt; &amp;'static FileDescriptorProto</pre></section>
     <section id='search' class="content hidden"></section>
