+++ docs/rusty_secrets/share_data/index.html
@@ -47,7 +47,7 @@
                    <a id="toggle-all-docs" href="javascript:void(0)" title="collapse all docs">
                        [<span class='inner'>&#x2212;</span>]
                    </a>
-               </span><a id='src-310' class='srclink' href='../../src/rusty_secrets/src/share_data.rs.html#5-345' title='goto source code'>[src]</a></span></h1>
+               </span><a id='src-312' class='srclink' href='../../src/rusty_secrets/src/share_data.rs.html#5-345' title='goto source code'>[src]</a></span></h1>
 <h2 id='structs' class='section-header'><a href="#structs">Structs</a></h2>
 <table>
                        <tr class=' module-item'>
