+++ docs/search-index.js
@@ -1,3 +1,3 @@
 var searchIndex = {};
-searchIndex["rusty_secrets"] = {"doc":"","items":[[5,"generate_shares","rusty_secrets","Performs threshold k-out-of-n Shamir secret sharing.",null,null],[5,"recover_secret","","Recovers the secret from a k-out-of-n Shamir secret sharing.",null,{"inputs":[{"name":"vec"},{"name":"bool"}],"output":{"name":"result"}}],[0,"custom_error","","",null,null],[3,"Error","rusty_secrets::custom_error","Error struct used for generating an `io::Error` from a generic description.",null,null],[5,"other_io_err","","Returns an `io::Error` from description string and optional detail string.\nParticularly useful in `Result` expressions.",null,{"inputs":[{"name":"str"},{"name":"option"}],"output":{"name":"error"}}],[5,"pie2io","","maps a `ParseIntError` to an `io::Error`",null,{"inputs":[{"name":"parseinterror"}],"output":{"name":"error"}}],[11,"fmt","","",0,null],[11,"new","","",0,{"inputs":[{"name":"str"},{"name":"option"}],"output":{"name":"error"}}],[11,"fmt","","",0,null],[11,"description","","",0,null],[11,"cause","","",0,null],[0,"share_data","rusty_secrets","",null,null],[3,"ShareData","rusty_secrets::share_data","",null,null],[5,"file_descriptor_proto","","",null,{"inputs":[],"output":{"name":"filedescriptorproto"}}],[11,"clone","","",1,null],[11,"default","","",1,{"inputs":[],"output":{"name":"sharedata"}}],[11,"new","","",1,{"inputs":[],"output":{"name":"sharedata"}}],[11,"default_instance","","",1,{"inputs":[],"output":{"name":"sharedata"}}],[11,"clear_shamir_data","","",1,null],[11,"has_shamir_data","","",1,null],[11,"set_shamir_data","","",1,null],[11,"mut_shamir_data","","",1,null],[11,"take_shamir_data","","",1,null],[11,"get_shamir_data","","",1,null],[11,"clear_signature","","",1,null],[11,"set_signature","","",1,null],[11,"mut_signature","","",1,null],[11,"take_signature","","",1,null],[11,"get_signature","","",1,null],[11,"clear_proof","","",1,null],[11,"has_proof","","",1,null],[11,"set_proof","","",1,null],[11,"mut_proof","","",1,null],[11,"take_proof","","",1,null],[11,"get_proof","","",1,null],[11,"is_initialized","","",1,null],[11,"merge_from","","",1,null],[11,"compute_size","","",1,null],[11,"write_to_with_cached_sizes","","",1,null],[11,"get_cached_size","","",1,null],[11,"get_unknown_fields","","",1,null],[11,"mut_unknown_fields","","",1,null],[11,"type_id","","",1,null],[11,"as_any","","",1,null],[11,"descriptor","","",1,null],[11,"new","","",1,{"inputs":[],"output":{"name":"sharedata"}}],[11,"descriptor_static","","",1,{"inputs":[{"name":"option"}],"output":{"name":"messagedescriptor"}}],[11,"clear","","",1,null],[11,"eq","","",1,null],[11,"fmt","","",1,null]],"paths":[[3,"Error"],[3,"ShareData"]]};
+searchIndex["rusty_secrets"] = {"doc":"`RustySecrets` implements Shamir Secret Sharing in Rust. It provides the possibility to sign shares.","items":[[5,"generate_shares","rusty_secrets","Performs threshold k-out-of-n Shamir secret sharing.",null,null],[5,"recover_secret","","Recovers the secret from a k-out-of-n Shamir secret sharing.",null,{"inputs":[{"name":"vec"},{"name":"bool"}],"output":{"name":"result"}}],[0,"share_data","","",null,null],[3,"ShareData","rusty_secrets::share_data","",null,null],[5,"file_descriptor_proto","","",null,{"inputs":[],"output":{"name":"filedescriptorproto"}}],[11,"clone","","",0,null],[11,"default","","",0,{"inputs":[],"output":{"name":"sharedata"}}],[11,"new","","",0,{"inputs":[],"output":{"name":"sharedata"}}],[11,"default_instance","","",0,{"inputs":[],"output":{"name":"sharedata"}}],[11,"clear_shamir_data","","",0,null],[11,"has_shamir_data","","",0,null],[11,"set_shamir_data","","",0,null],[11,"mut_shamir_data","","",0,null],[11,"take_shamir_data","","",0,null],[11,"get_shamir_data","","",0,null],[11,"clear_signature","","",0,null],[11,"set_signature","","",0,null],[11,"mut_signature","","",0,null],[11,"take_signature","","",0,null],[11,"get_signature","","",0,null],[11,"clear_proof","","",0,null],[11,"has_proof","","",0,null],[11,"set_proof","","",0,null],[11,"mut_proof","","",0,null],[11,"take_proof","","",0,null],[11,"get_proof","","",0,null],[11,"is_initialized","","",0,null],[11,"merge_from","","",0,null],[11,"compute_size","","",0,null],[11,"write_to_with_cached_sizes","","",0,null],[11,"get_cached_size","","",0,null],[11,"get_unknown_fields","","",0,null],[11,"mut_unknown_fields","","",0,null],[11,"type_id","","",0,null],[11,"as_any","","",0,null],[11,"descriptor","","",0,null],[11,"new","","",0,{"inputs":[],"output":{"name":"sharedata"}}],[11,"descriptor_static","","",0,{"inputs":[{"name":"option"}],"output":{"name":"messagedescriptor"}}],[11,"clear","","",0,null],[11,"eq","","",0,null],[11,"fmt","","",0,null]],"paths":[[3,"ShareData"]]};
 initSearch(searchIndex);
