+++ docs/src/rusty_secrets/src/lib.rs.html
@@ -70,14 +70,23 @@
 <span id="26">26</span>
 <span id="27">27</span>
 <span id="28">28</span>
+<span id="29">29</span>
+<span id="30">30</span>
+<span id="31">31</span>
+<span id="32">32</span>
+<span id="33">33</span>
+<span id="34">34</span>
+<span id="35">35</span>
 </pre><pre class='rust '>
 <span class='attribute'>#<span class='op'>!</span>[<span class='ident'>deny</span>(
-    <span class='comment'>//missing_docs,</span>
+    <span class='ident'>missing_docs</span>,
     <span class='ident'>missing_debug_implementations</span>, <span class='ident'>missing_copy_implementations</span>,
     <span class='ident'>trivial_casts</span>, <span class='ident'>trivial_numeric_casts</span>,
     <span class='ident'>unsafe_code</span>, <span class='ident'>unstable_features</span>,
     <span class='ident'>unused_import_braces</span>, <span class='ident'>unused_qualifications</span>
 )]</span>
+
+<span class='doccomment'>//! `RustySecrets` implements Shamir Secret Sharing in Rust. It provides the possibility to sign shares.</span>
 
 <span class='kw'>extern</span> <span class='kw'>crate</span> <span class='ident'>protobuf</span>;
 <span class='kw'>extern</span> <span class='kw'>crate</span> <span class='ident'>rustc_serialize</span> <span class='kw'>as</span> <span class='ident'>serialize</span>;
@@ -85,20 +94,25 @@
 <span class='kw'>extern</span> <span class='kw'>crate</span> <span class='ident'>merkle_sigs</span>;
 <span class='kw'>extern</span> <span class='kw'>crate</span> <span class='ident'>ring</span>;
 
-<span class='kw'>mod</span> <span class='ident'>gf256</span>;
-<span class='kw'>mod</span> <span class='ident'>validation</span>;
-<span class='kw'>mod</span> <span class='ident'>interpolation</span>;
-<span class='kw'>mod</span> <span class='ident'>share_format</span>;
-<span class='kw'>mod</span> <span class='ident'>sss</span>;
-
-<span class='kw'>pub</span> <span class='kw'>mod</span> <span class='ident'>custom_error</span>;
-<span class='kw'>pub</span> <span class='kw'>use</span> <span class='ident'>sss</span>::<span class='ident'>generate_shares</span>;
-<span class='kw'>pub</span> <span class='kw'>use</span> <span class='ident'>sss</span>::<span class='ident'>recover_secret</span>;
-<span class='kw'>pub</span> <span class='kw'>mod</span> <span class='ident'>share_data</span>;
-
 <span class='kw'>use</span> <span class='ident'>ring</span>::<span class='ident'>digest</span>::{<span class='ident'>Algorithm</span>, <span class='ident'>SHA512</span>};
 <span class='attribute'>#[<span class='ident'>allow</span>(<span class='ident'>non_upper_case_globals</span>)]</span>
 <span class='kw'>static</span> <span class='ident'>digest</span>: <span class='kw-2'>&amp;</span><span class='lifetime'>&#39;static</span> <span class='ident'>Algorithm</span> <span class='op'>=</span> <span class='kw-2'>&amp;</span><span class='ident'>SHA512</span>;
+
+<span class='kw'>mod</span> <span class='ident'>custom_error</span>;
+<span class='kw'>mod</span> <span class='ident'>gf256</span>;
+<span class='kw'>mod</span> <span class='ident'>interpolation</span>;
+<span class='kw'>mod</span> <span class='ident'>share_format</span>;
+<span class='kw'>mod</span> <span class='ident'>sss</span>;
+<span class='kw'>mod</span> <span class='ident'>validation</span>;
+
+<span class='kw'>pub</span> <span class='kw'>use</span> <span class='ident'>sss</span>::<span class='ident'>generate_shares</span>;
+<span class='kw'>pub</span> <span class='kw'>use</span> <span class='ident'>sss</span>::<span class='ident'>recover_secret</span>;
+
+<span class='attribute'>#[<span class='ident'>allow</span>(<span class='ident'>missing_docs</span>)]</span>
+<span class='kw'>pub</span> <span class='kw'>mod</span> <span class='ident'>share_data</span>;
+
+<span class='attribute'>#[<span class='ident'>cfg</span>(<span class='ident'>test</span>)]</span>
+<span class='kw'>mod</span> <span class='ident'>tests</span>;
 </pre>
 </section>
     <section id='search' class="content hidden"></section>
